api update

+++ src/pages/api/delete-group.ts
@@ -1,16 +1,28 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth";
 import prisma from "services/prisma";
 import { prismaErrHandler } from "services/prismaErrHandler";
+import { authOptions } from "./auth/[...nextauth]";
+import { getGroupPerm } from "services/utils";
+import { TypedRequestBody } from "@/types";
 
 export default async function handle(
-  req: NextApiRequest,
+  req: TypedRequestBody<{
+    id: number;
+  }>,
   res: NextApiResponse
 ) {
+  //--recv--
   const { id } = req.body;
+  //--API Protection--
+  const session = await getServerSession(req, res, authOptions);
+  const groupPerm = getGroupPerm(session?.user, id);
+  if (groupPerm < 1) return res.status(403).json("Forbidden");
+  //--operation--
   try {
     const op = await prisma.group.delete({
       where: {
-        id: id,
+        id,
       },
     });
     return res.status(200).json(op);
