+++ src/pages/api/delete-item.ts
@@ -1,16 +1,29 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth";
 import prisma from "services/prisma";
 import { prismaErrHandler } from "services/prismaErrHandler";
+import { authOptions } from "./auth/[...nextauth]";
+import { getGroupPerm } from "services/utils";
+import { TypedRequestBody } from "@/types";
 
 export default async function handle(
-  req: NextApiRequest,
+  req: TypedRequestBody<{
+    groupId: number;
+    itemId: number;
+  }>,
   res: NextApiResponse
 ) {
-  const { id } = req.body;
+  //--rcv--
+  const { groupId, itemId } = req.body;
+  //--API Protection--
+  const session = await getServerSession(req, res, authOptions);
+  const groupPerm = getGroupPerm(session?.user, groupId);
+  if (groupPerm < 1) return res.status(403).json("Forbidden");
+  //--operation--
   try {
     const op = await prisma.item.delete({
       where: {
-        id: id,
+        id: itemId,
       },
     });
     return res.status(200).json(op);
